Changes to unavailable volumes output
Made changes to __init__.py file in volumes folder. Changes are to the grep function and will search the ERROR.err file and add volume ids found there to a list. If the length of the list is less than 100, it will display the unavailable volumes in the terminal and also save them to a file named "volume_na.txt" to the users home directory. If the length of the list equals 100, then an error message is displayed in the terminal.