Changed -s flag default
Changed the function of the -s flag so that if a user does NOT call the flag in conjunction with the -hf or -hfc flags the .csv files containing the removed headers/footers will be saved to the output directory by default. If a user does call the -s flag in conjunction with the -hf or -hfc flags then the .csv files will be "skipped" and not saved to the output directory.